Do not delete role viewer on user while updating group

+++ src/main/java/nl/opengeogroep/safetymaps/server/admin/stripes/EditGroupsActionBean.java
@@ -272,7 +272,7 @@ public class EditGroupsActionBean implements ActionBean, ValidationErrorHandler 
         }
 
         for(String user : users) {
-            qr().update("delete from " + USER_ROLE_TABLE + " where username = ? and role != ?", user, role);
+            qr().update("delete from " + USER_ROLE_TABLE + " where username = ? and role != ? and role != 'viewer'", user, role);
         }
 
         for(String user : users) {
